+++ src/Application/Invitation/InvitationException.php
@@ -2,12 +2,18 @@
 
 namespace EventFlow\Application\Invitation;
 
+use EventFlow\Application\Error\ControlledFailure;
 use RuntimeException;
 
-final class InvitationException extends RuntimeException
+final class InvitationException extends RuntimeException implements ControlledFailure
 {
     public function __construct(public readonly string $safeCode)
     {
         parent::__construct($safeCode);
+    }
+
+    public function safeCode(): string
+    {
+        return $this->safeCode;
     }
 }